Stop retaining old authids and empty sets in router.
When all of the given sessions had disconnected for a given authid, the authid and empty set were
being held in a map. This meant that when Anonymous Authentication was used, you could have lots of
authids and empty sets being kept around, but the authid would be unlikely to be used again. This
commit removes the authid and empty set from the map, allowing that memory to be garbage collected.