Fix handling YIELD and ERROR of unknown invocation
When they are applied to an unknown invocation, it doesn't necessarily
mean that they are protocol errors. This could happen due to a race
condition.

For a possible scenario, see issue #1563.

And even if it was a protocol error, e.g. it is ending an invocation of
a different session, then a different check kicks in and throws a
ProtocolError exception anyway.

This partly fixes #1563