Partly fix YIELD/ERROR race (#1592)
* Remove duplicate keys in config schema

* Add vscode settings file

* Add prepareUbuntu make target

I had a hard time finding out what went wrong running make test. I
could put in the docs how I fixed it but figgered I could also add a
make target that fixes it on Ubuntu.

* Fix handling YIELD and ERROR of unknown invocation

When they are applied to an unknown invocation, it doesn't necessarily
mean that they are protocol errors. This could happen due to a race
condition.

For a possible scenario, see issue #1563.

And even if it was a protocol error, e.g. it is ending an invocation of
a different session, then a different check kicks in and throws a
ProtocolError exception anyway.

This partly fixes #1563