common/docs: allow TCP user timeout configuration in endpoint
TCP_USER_TIMEOUT (see `man tcp`) is a socket option that causes TCP
connections to be closed if transmitted data is not acked within the
given time (in milliseconds). This allows failing fast instead of waiting
for ACKs to arrive up to 20 min on dead connections.

A use case for this is an embedded device connecting to crossbar. If
this is shut down forcibly, the connection hangs and subsequent
procedure registrations fail until the connection is closed. This can
take up to 20 minutes.

Signed-off-by: Bastian Krause <bst@pengutronix.de>