Allow TCP user timeout configuration in endpoint (#1930)
* common: twisted: sharedport: rename SharedPort -> CustomTCPPort, SharedTLSPort -> CustomTCPTLSPort

These ports have the potential to be used pretty generic. In a future
commit, they will gain other functionality, so name them accordingly
already.

Leave the name of the module as it is to keep the caused mess small.

Signed-off-by: Bastian Krause <bst@pengutronix.de>

* common: twisted: sharedport: allow setting TCP user timeout in subclassed TCP ports

Signed-off-by: Bastian Krause <bst@pengutronix.de>

* common: twisted: endpoint: read the shared config attribute once

We need to check the attribute value anyway, so read it into a handy
variable right away. A future commit will add another variable to check,
so we prepare to keep the check readable.

Signed-off-by: Bastian Krause <bst@pengutronix.de>

* common/docs: allow TCP user timeout configuration in endpoint

TCP_USER_TIMEOUT (see `man tcp`) is a socket option that causes TCP
connections to be closed if transmitted data is not acked within the
given time (in milliseconds). This allows failing fast instead of waiting
for ACKs to arrive up to 20 min on dead connections.

A use case for this is an embedded device connecting to crossbar. If
this is shut down forcibly, the connection hangs and subsequent
procedure registrations fail until the connection is closed. This can
take up to 20 minutes.

Signed-off-by: Bastian Krause <bst@pengutronix.de>